New Grind Rail System
New grind rail system that uses a timeline and play rate change to simulate/fake physics.
Still needs tweaking and for the player to maintain their speed when exiting the rail, as well as allowing for jumping off the rail. Overall, still works so much better.
Need to remove old rail system as well.