Overhauled Spindash Charge System
The old Adventure/Classic Spindash charge systems are removed, in their place is a unified system that allows for more flexibility in options. Gradual charge, jump rev, requiring the player to be at a standstill to charge, and more are customizable. Various bugs, such as the charge/roll sounds not playing, have also been fixed with the new system.
Added the ability for Dash Panels (if toggled in the actor) to put the player into the Roll state on contact.
Removed E_SpindashType, as this asset has been made obsolete with the system change for Spindash charging.