Edits
Animation for using the Thunder Shield's shield skill has been exaggerated slightly.
If the player has the Crouch skill, they will now crouch first before charging a Spindash (which must be done through a jump rev).
The player will now maintain their roll state when under geometry, similar to the Slide. They'll be thrusted forward whenever they go below the roll's speed threshold while also under geometry.
Added option for Spindash releases to use either AddImpulse (additive speed) or SetVelocity (fixed speed).
Fixed an issue where changing the capsule half height for the AdventureSonicPawn would result in Sonic floating above the floor, even while crouched/rolling. This was fixed by setting the capsule component's scale to 1;1;1, as it was strangely 1.2;1.2;1.2 originally.
Fixed an issue where the Thunder Shield would not properly magnetize Rings. This was due to an improper check that looked at if the RingActor had LightdashRing set to true, resulting in some Ring patches within NewTestMap to not be properly magnetized. The check now looks against if ShowLightdashParticle is true or not, as these are Rings the shield should not magnetize (the particle conveys to the player it's clearly a lightdash trail).
Fixed an improper check when magnetizing Rings where the player having a Thunder Shield wouldn't be factored into the magnetization check. With this fixed, Rings can be magnetized to the player from a further distance.
Fixed an issue where the player could use the Lightdash on a magnetized Ring (either through boosting or a Thunder Shield) by forcing LightdashRing to be false within the RingActor when magnetizing.
New geometry added to NewTestMap.
Possibly other smaller edits I forgot to jot down.