Sliding Slope Physics + Jump Action Edits
Sliding now uses the same function that rolling uses for slope physics rather than a fixed velocity cutdown, allowing it to be used down and up slopes properly. Sliding uses a separate Vector 2D variable for slope influence.
Fixed an issue where the feedback for a Jumpdash would not properly execute outside of jumping then jumpdashing. This was solved by connecting the NotValid pin on IsValid (TempJumpball) to the rest of the Jumpdash code.
Fixed an issue where the particle for a charged Dropdash would persist if the player completed the Dropdash charge right as they landed on the ground. This was solved by stopping the DropdashCharge timeline in the Dropdash skill's graph on the OnLanded event.