Pulley Edits
Fixed an issue where errors would be thrown up with the PulleyActor not finding a value for the Player input on the Retract event called if StartRetracted is true. Solved by using UpdatePulley on BeginPlay rather than the Retract/Deploy events.
Fixed an issue where a pulley that was retracted could still be grabbed on to. This occurs because the pulley always assumes it's deployed and never checks against a variable when the player tries to grab it. Solved by adding a runtime "Retracted" boolean to the PulleyActor that saves if the pulley is retracted or not.
Pulleys that are not of the SA2 type will now put the player into the Spring animation state rather than PulleyJump. This animation was by default the Spring animation anyways.