Attempted Slope Exit Fix
Attempted to fix an issue where the player would not be at the correct rotation when exiting a straight vertical wall in slope physics. This occurs likely because the PawnOrientation function rotates the player to match their velocity, which is straight vertical when exiting the vertical wall and thus orients the player perpendicular to where they should be facing. No known solution has been found for this outside of having geometry that angles the player more horizontal before exiting the slope, which is not a very reliable solution.
Cleaned up some graphs in the pawn, primarily those related to camera changes.