Rail Switching
Added Rail Switching back in with a new system more similar to SA2/Heroes than the Boost games. Sonic is detached from the current rail and launched to his left/right diagonal forward.
Added a DoOnce node to NewRailDetect in DashPlayerPawn. This somewhat solves the issue of entering a grind rail resulting in a short game freeze (freeze is longer and more crash prone the longer the rail).