FinishLevel Hotfix + Minor Edits
Fixed an issue with saving data in the FinishLevel function. Done by correctly making the SaveRing graph save Rings and not Score.
Removed the Grinding physics mode from E_PhysMode and the matching struct in DashPlayerPawn.
Tried to allow proper jumping off of enemies instead of always defaulting to a trick.
Disconnected the Slowmo event's code.