Zipline Cam Hotfix
Fixed an issue where Ziplines would always function as if CameraLock was true. This was because of how the CameraMode graph in the tick function GroundMovement worked. Solved the issue by adding a check for NOT being in the Pulley animation state within the branch before setting the camera mode to LoopingCamMode.