Credits Scene Started
New Credits Scene.
Plan currently is to have the player be able to run around the scene and collect Rings and other goodies as the game's credits roll. The player can hold Pause to skip the credits and return to the main menu at any time. The player will also be able to find and collect all 7 Chaos Emeralds to showcase the new emerald saving system (which will work similar to Red Rings, but be a persistent save to PlayerSave rather than per level, system change is done to prevent having more than 7 emeralds at a time).