Attack State Fix
Fixed an issue where Boosting, then Sliding, then stopping a Slide while still Boosting results in an incorrect Attack State (Normal when it should be Boost).
Fixed an issue where boosting into a grind rail stops the Boost. Done by adding a unique bool input to StopActions (which is called within the StartGrinding graph of NewRailDetect) that, if true, stops the Boost. This is set to false exclusively in NewRailDetect.