Fatal Crash Hotfix
Sort of temporary hotfix to a fatal crash that occurs when SonicAnimBP transitions from IdleLoop to LongIdleLoopStart. The AnimBP just uses a non-connected duplicate of the IdleLoop state, which somehow fixes the issue.

For those reading, the exact error in the editor crash window is an Assertion Failed error and has to do with array indexes and the given size being -1. I have no idea what this means.