Removed ShowLv System
Removed the Show boolean from STR_LvSaveData, as this was resulting in an issue where newly added entries to the level data table weren't appearing (old SG didn't have given array entry).
The workaround was to delete the SG asset in the project files to let everything refresh/reset, but this is cumbersome to do for every new level you add.

TLDR; you can add levels and it be much more straightforward.